requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib
pillow
xgboost
